src/requirements/base.txt
asgiref==3.8.1
Django==4.2.9
django-environ==0.10.0
djangorestframework==3.15.2
sqlparse==0.5.0
typing_extensions==4.12.2
boto3==1.34.141
celery==5.4.0
redis==5.0.7
django-cors-headers==4.4.0
django-autoslug==1.9.9
drf-yasg==1.21.5
pillow==9.4.0
argon2-cffi==21.3.0
requests==2.31.0
